Docker_dir/camera_create/requirements.txt
python-telegram-bot==13.4.1
pymongo==3.11.2
dnspython==2.0.0
schedule==1.0.0
sockets